refactor: Auth 도메인 관련 파일 경로 수정
